+++ src/Typing.js
@@ -1,0 +1,13 @@
+import React, { Component, Fragment } from 'react';
+
+class Typing extends Component {
+  render() {
+    return (
+      <Fragment>
+        <p>타이핑 애니메이션이 실행되는 부분입니다.</p>
+      </Fragment>
+    );
+  }
+}
+
+export default Typing;
